// frontend/src/app/models/employee.ts
export interface Employee {
  id?: number;
  name: string;
  department: string;
  jobRole: string;
  age: number;
  attrition?: string;
  businessTravel?: string;
  dailyRate?: number;
  distanceFromHome?: number;
  education?: number;
  educationField?: string;
  employeeCount?: number;
  employeeNumber?: number;
  environmentSatisfaction?: number;
  gender?: string;
  hourlyRate?: number;
  jobInvolvement?: number;
  jobLevel?: number;
  jobSatisfaction?: number;
  maritalStatus?: string;
  monthlyIncome?: number;
  monthlyRate?: number;
  numCompaniesWorked?: number;
  over18?: string;
  overTime?: string;
  percentSalaryHike?: number;
  performanceRating?: number;
  relationshipSatisfaction?: number;
  standardHours?: number;
  stockOptionLevel?: number;
  totalWorkingYears?: number;
  trainingTimesLastYear?: number;
  workLifeBalance?: number;
  yearsAtCompany?: number;
  yearsInCurrentRole?: number;
  yearsSinceLastPromotion?: number;
  yearsWithCurrManager?: number;
}